refactor(card-editor): replace bulk editor with single-inject controller
