+++ src/views/deck/card-editor/list.vue
@@ -1,18 +1,14 @@
 <script setup lang="ts">
 import ListItem from './list-item.vue'
 import { inject, useTemplateRef } from 'vue'
-import { type CardBulkEditor } from '@/composables/card-bulk-editor'
-import { useInfiniteScroll } from '@/composables/use-infinite-scroll'
-import { useCardsInDeckInfiniteQuery } from '@/api/cards'
+import { type CardListController } from '@/composables/card-list-controller'
 
-const { all_cards, getKey } = inject<CardBulkEditor>('card-editor')!
-const cards_query = inject<ReturnType<typeof useCardsInDeckInfiniteQuery>>('cards-query')!
+const { all_cards, getKey, hasNextPage, isLoading, observeSentinel } =
+  inject<CardListController>('card-editor')!
 
 const sentinel = useTemplateRef<HTMLElement>('sentinel')
 
-useInfiniteScroll(sentinel, () => cards_query.loadNextPage(), {
-  enabled: () => cards_query.hasNextPage.value && !cards_query.isLoading.value
-})
+observeSentinel(sentinel)
 </script>
 
 <template>
@@ -26,12 +22,12 @@ useInfiniteScroll(sentinel, () => cards_query.loadNextPage(), {
     >
     </list-item>
     <div
-      v-if="cards_query.hasNextPage.value"
+      v-if="hasNextPage"
       ref="sentinel"
       data-testid="card-list__sentinel"
       class="w-full py-6 flex items-center justify-center text-brown-500"
     >
-      <span v-if="cards_query.isLoading.value">Loading…</span>
+      <span v-if="isLoading">Loading…</span>
     </div>
     <slot></slot>
   </div>
