+++ src/views/deck/card-importer.vue
@@ -1,8 +1,9 @@
 <script setup lang="ts">
-import { computed, ref } from 'vue'
+import { computed, inject, ref } from 'vue'
 import UiButton from '@/components/ui-kit/button.vue'
 import Card from '@/components/card/index.vue'
 import { useBulkInsertCardsInDeckMutation } from '@/api/cards'
+import { type CardListController } from '@/composables/card-list-controller'
 
 type CardDraft = { front_text: string; back_text: string }
 
@@ -12,7 +13,7 @@ const raw_text = ref<string>('')
 const cards = ref<CardDraft[]>([])
 const bulk_insert_mutation = useBulkInsertCardsInDeckMutation()
 
-const { deck_id } = defineProps<{ deck_id: number }>()
+const { deck_id } = inject<CardListController>('card-editor')!
 
 const has_unsaved_changes = computed(() => cards.value.length > 0)
 
@@ -30,7 +31,8 @@ async function onSave() {
   if (!has_unsaved_changes.value || saving.value) return
 
   saving.value = true
-  await bulk_insert_mutation.mutateAsync({ deck_id, cards: cards.value })
+  if (deck_id.value === undefined) return
+  await bulk_insert_mutation.mutateAsync({ deck_id: deck_id.value, cards: cards.value })
   saving.value = false
   cards.value = []
 }
